pyproject.toml
[tool.poetry]
name = "gpt-rag"
version = "0.1.0"
description = "local gpt client for RAG"
authors = ["Robert Hirsch <dev@robert-hirsch.de>"]
license = "GPLv3"
readme = "README.md"
packages = [{include = "gpt_rag", from = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
uvicorn = "^0.30.1"
fastapi = "^0.111.0"
sqlalchemy = "^2.0.30"
pwdlib = {extras = ["argon2"], version = "^0.2.0"}


itsdangerous = "^2.2.0"
websockets = "^11.0"
gradio = "^4.36.1"
openai = "^1.34.0"
python-dotenv = "^1.0.1"
[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
gpt-rag = "main:main"